+++ lucene/core/src/java/org/apache/lucene/search/similarities/AxiomaticF3LOG.java
@@ -17,10 +17,10 @@
 package org.apache.lucene.search.similarities;
 
 /**
- * F2EXP is defined as Sum(tf(term_doc_freq)*IDF(term)-gamma(docLen, queryLen))
+ * F3EXP is defined as Sum(tf(term_doc_freq)*IDF(term)-gamma(docLen, queryLen))
  * where IDF(t) = ln((N+1)/df(t)) N=total num of docs, df=doc freq
  * gamma(docLen, queryLen) = (docLen-queryLen)*queryLen*s/avdl
- *
+ * NOTE: the gamma function of this similarity creates negative scores
  * @lucene.experimental
  */
 public class AxiomaticF3LOG extends Axiomatic {
@@ -44,40 +44,40 @@ public class AxiomaticF3LOG extends Axiomatic {
    * compute the term frequency component
    */
   @Override
-  protected float tf(BasicStats stats, float freq, float docLen) {
-    if (freq <= 0.0) return 0f;
-    return (float) (1 + Math.log(1 + Math.log(freq)));
+  protected double tf(BasicStats stats, double freq, double docLen) {
+    if (freq <= 0.0) return 0.0;
+    return 1 + Math.log(1 + Math.log(freq));
   }
 
   /**
    * compute the document length component
    */
   @Override
-  protected float ln(BasicStats stats, float freq, float docLen) {
-    return 1f;
+  protected double ln(BasicStats stats, double freq, double docLen) {
+    return 1.0;
   }
 
   /**
    * compute the mixed term frequency and document length component
    */
   @Override
-  protected float tfln(BasicStats stats, float freq, float docLen) {
-    return 1f;
+  protected double tfln(BasicStats stats, double freq, double docLen) {
+    return 1.0;
   }
 
   /**
    * compute the inverted document frequency component
    */
   @Override
-  protected float idf(BasicStats stats, float freq, float docLen) {
-    return (float) Math.log((stats.getNumberOfDocuments() + 1.0) / stats.getDocFreq());
+  protected double idf(BasicStats stats, double freq, double docLen) {
+    return Math.log((stats.getNumberOfDocuments() + 1.0) / stats.getDocFreq());
   }
 
   /**
    * compute the gamma component
    */
   @Override
-  protected float gamma(BasicStats stats, float freq, float docLen) {
+  protected double gamma(BasicStats stats, double freq, double docLen) {
     return (docLen - this.queryLen) * this.s * this.queryLen / stats.getAvgFieldLength();
   }
 }
